+++ frontend/src/app/admin/manage-product/promotion-product/promotion-product.component.ts
@@ -1,11 +1,11 @@
-import { Component, OnInit } from '@angular/core';
-import { Product } from '../../../model/product.model';
-import { ProductService } from '../../../service/product-service.service';
-import { RouterLink } from '@angular/router';
-import { ImageService } from '../../../service/img-service.service';
-import { CommonModule } from '@angular/common';
-import { Promotion } from '../../../model/promotion.model';
-import { PromotionService } from '../../../service/promotion-service.service';
+import {Component, OnInit} from '@angular/core';
+import {Product} from '../../../model/product.model';
+import {ProductService} from '../../../service/product-service.service';
+import {RouterLink} from '@angular/router';
+import {ImageService} from '../../../service/img-service.service';
+import {CommonModule} from '@angular/common';
+import {Promotion} from '../../../model/promotion.model';
+import {PromotionService} from '../../../service/promotion-service.service';
 
 @Component({
   selector: 'app-promotion-product',
@@ -18,12 +18,14 @@ export class PromotionProductComponent implements OnInit {
   product: any;
 
   constructor(private productService: ProductService,
-    private imgService: ImageService,
-    private promotionService: PromotionService,) { }
+              private imgService: ImageService,
+              private promotionService: PromotionService,) {
+  }
 
   imgAvatars: { [key: string]: string } = {};
   products: Product[] = []
   promotion: Promotion = new Promotion
+
   ngOnInit(): void {
     this.getAllwithPromotion()
   }
@@ -42,20 +44,23 @@ export class PromotionProductComponent implements OnInit {
     if (imageName) {
       this.imgService.getImage(imageName).subscribe(
         (data: any) => {
-          const blob = new Blob([data], { type: 'image/*' });
+          const blob = new Blob([data], {type: 'image/*'});
           this.imgAvatars[product_id] = URL.createObjectURL(blob);
         },
         error => {
           console.log(error);
         });
-    } else { }
+    } else {
+    }
   }
 
-  delete(promotion_id:string) {
-    this.promotionService.deleteById(promotion_id).subscribe(
-      (data: any) => {
-        this.getAllwithPromotion();
-      });
+  delete(promotion_id: string) {
+    if (window.confirm("Are you sure want to delete this promotion ?")) {
+      this.promotionService.deleteById(promotion_id).subscribe(
+        (data: any) => {
+          this.getAllwithPromotion();
+        });
+    }
   }
 }
